adapter: move statement logging into adapter crate
Add some helper From methods.
